+++ packages/analytics-js-cookies/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.5.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.5.0...@rudderstack/analytics-js-cookies@0.5.1) (2025-08-11)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.23.0`
 ## [0.5.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.4.27...@rudderstack/analytics-js-cookies@0.5.0) (2025-07-21)
 
 ### Dependency Updates