+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.52.2](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.1...rudder-sdk-js@2.52.2) (2025-08-11)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.16.1`
 ## [2.52.1](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.0...rudder-sdk-js@2.52.1) (2025-07-30)
 
 ### Dependency Updates